SyntheticLikelihood

Documentation for SyntheticLikelihood package. Simulator functions should have a single positional argument of parameters (and can have arbitrary key word arguments). Summary functions should take the output from the simulator directly as a single positional argument (and can have arbitrary key word arguments).

Function list

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.synthetic_likelihood

Function descriptions

SyntheticLikelihood.simulate_n_s — Method

Simulates summary statistics from the model under a fixed parameter vector, or an array of parameter vectors. If a vector is used, n_sim is specified as the number of simulations. If an array is used, one simulation is carried out using each row as a parameter vector.

simulate_n_s(θ; simulator, summary, n_sim, simulator_kwargs, summary_kwargs)

Arguments

θ::Vector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator.

n_sim::Int Number of simulations.

simulator_kwargs Kwargs passed to simulator.

summary_kwargs Kwargs passed to summary.

SyntheticLikelihood.simulate_n_s — Method

As for above, but uses a matrix of parameter values, carrying out one simulation from each row of θ.

simulate_n_s(θ; simulator, summary, simulator_kwargs, summary_kwargs)

SyntheticLikelihood.synthetic_likelihood — Method

Evaluates synthetic likelhood of observed data for a fixed parameter vector using a multivariate Gaussian assumption as in (Simon Wood, 2010).

synthetic_likelihood(θ; simulator, summary, s_true, n_sim, simulator_kwargs, summary_kwargs)

Arguments

θ::Vector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator.

s_true::Vector Observed summary statistics.

n_sim::Int Number of simulations to use.

simulator_kwargs Kwargs passed to simulator.

summary_kwargs Kwargs passed to summary.