Documentation

Sampling

SyntheticLikelihood.RiemannianULA — Type

Sampler object for Riemannian ULA. Uses the update: θ := θ - ϵ²H⁻¹*∇ - ϵ√H⁻¹ z, where z ∼ N(0, I).

step_size

SyntheticLikelihood.ULA — Type

Sampler for unadjusted langevin algorithm. Uses a discrete time Euler approximation of the langevin diffusion, given by the update θ := θ - η/2 * ∇ + ξ, where ξ is given by N(0, ηI).

step_size

SyntheticLikelihood.run_sampler! — Method

Run the sampling algorithm. Data to collect at each iteration is specified by collect_data, and should be a subset of [:θ, :objective, :gradient, :hessian, :counter].

run_sampler!(sampler, local_approximation; init_θ, n_steps, collect_data)

Returns a tuple, with keys matching collect_data.

Implementation details

To implement a new sampler:

Create a subtype of AbstractSampler, which takes the hyperparameters.

Create a subtype of AbstractSamplerState which stores the state of the sampler at each iteration.

Implement the get_init_state and update! functions to update the state

(e.g. parameters, gradients, objective function value).

Objectives

SyntheticLikelihood.LocalLikelihood — Type

Contains the hyperparameters for getting a local approximation of the negative log-likelihood surface using local regressions.

simulator

summary

s_true

P

Initial distribution used to peturb the parameter value.

n_sim

The number of peturbed points to use for the local regression.

P_regularizer

Method to regularise the inverse Hessian to a reasonable covariance matrix.

valid_params

Parameter constraints. Function that takes a parameter vector and returns true (for valid parameters) or false (for invalid parameters). Defualt is θ -> true

SyntheticLikelihood.LocalPosterior — Type

Contains the hyperparameters for getting a local approximation of the posterior. In contrast to the likelihood version, a prior is provided, P is by defualt a MvNormal with covariance 0.5*cov(prior), and valid_params checks whether proposed points fall within the prior support.

prior

Prior distribution (either multivariate or Product distribution)

simulator

summary

s_true

P

n_sim

P_regularizer

valid_params

Implementation details

To implement a new objective:

Create a subtype of LocalApproximation defining the hyperpameters.

Implement obj_grad_hess function to estimate the objective, gradient and Hessian.

Matrix regularisation

As the Hessian inferred from local regressions is not guarenteed to be positive definite, we may need to modify the Hessian (or its inverse) in some way.

SyntheticLikelihood.KitchenSink — Type

Carries out the following steps:

Use soft_abs to get absolute values of the eigenvalues.

If required, shrink the variances towards the reference.

Decompose the covariance into the variance and correlation.

Divide the correlation matrix by a constant (if reqiured) to reach condition threshold.

Set non diagonal correlation to zero if < τ

Reconstruct the covariance matrix.

See http://parker.ad.siu.edu/Olive/slch6.pdf for correlation regularization, and https://doi.org/10.1007/978-3-642-40020-9_35 for soft abs.

SyntheticLikelihood.regularize — Method

Regularise the covariance matrix.

Plotting

Some convenience functions for plotting:

SyntheticLikelihood.plot_prior_posterior_density — Method

Plot the marginal priors and posterior, along with the true parameter values if available.

plot_prior_posterior_density(prior, θ_samples; param_names)

Private/Unexported interface

SyntheticLikelihood.LocalΣ — Type

Struct that contains the estimated local properties of Σ (the covariance matrix of the summary statistics).

Σ

∂

Σ The (nₛ×nₛ) (estimated) covariance matrix of the summary statistics.

∂ The (nₛ×nₛ×n_θ) matrix of estimated first derivitives of Σ.

SyntheticLikelihood.Localμ — Type

Struct that contains the estimated local properties of μ (the expected values of the summary statistics).

Fields

μ::Float64 Means of the summary statistics.

∂::Vector{Float64} First derivitives w.r.t. parameters (nₛ×n_θ).

∂²::Matrix{Float64} Second derivitive w.r.t. parameters (nₛ×nθ×nθ).

ϵ::Vector{Float64} Residuals of predicted summary statistics (nₛ×nₛᵢₘ).

SyntheticLikelihood.ObjGradHess — Type

Struct for containing the objective function value, along with the gradient and hessian if appropriate (defualt to nothing).

objective

gradient

hessian

SyntheticLikelihood.ObjectSummaryLogger — Type

Used to store and print information about an object. Useful for printing neat debugging messages.

SyntheticLikelihood.add_state! — Method

Add data to the data tuple.

SyntheticLikelihood.analytic_mvn_posterior — Method

Gets the analytic posterior distribution from a normal prior and normal likelihood. Useful for testing.

SyntheticLikelihood.cor_to_cov — Method

Convert correlation matrix to covariance matrix.

SyntheticLikelihood.cov_logdet_reg — Method

Scale matrix to a particular log-determinant value.

SyntheticLikelihood.cov_to_cor — Method

Convert covariance matrix to correlation matrix. Returns tuple (R, σ²)

SyntheticLikelihood.deterministic_test_simulator — Method

Deterministic simulator useful for testing.

SyntheticLikelihood.glm_local_Σ — Method

Use a gamma distributed GLM with log link function to estimate the local properties of the covariance matrix of the statistics Σ. θ should not have a bias term (added internally).

Specifically, this function:

Creates a rough initial Σ estimate using cov(ϵ).

Estimates the diagonal elements Σⱼⱼ, and ∂Σⱼⱼ using local regression.

Esimates off-diagonal elements of Σ by scaling the sample correlation matrix with √Σⱼⱼ (standard deviations).

Esimate off-diagonal gradients ∂Σᵢⱼ by averaging the coefficients associated

with indices i and j.

glm_local_Σ(; θᵢ, θ, ϵ)

Arguments

θᵢ::AbstractVector Original θ (used for centering).

θ::AbstractMatrix Peturbed θ from local area.

ϵ::AbstractMatrix Residuals from quadratic regression (nsim × nsumstats).

SyntheticLikelihood.halve_update_until_valid — Method

Halve update term until valid proposal found (e.g. with non-zero prior density). Returns the modified update term.

halve_update_until_valid(Δ, θ, valid_params)

Arguments

θ Current parameter vector.

Δ Proposed update term.

valid_params Returns true if valid and false if invalid proposal.

SyntheticLikelihood.init_data_tuple — Method

Initialises a named tuple containing Vectors with undefined values. Used with samplers to store results. State just provides an "example" state from which to infer types of vectors in the array. Names of the named tuple are the symbols provided.

SyntheticLikelihood.likelihood_calc — Method

Use results from local regressions to estimate the negative log-likelihood function value, gradient and hessian.

SyntheticLikelihood.likelihood_obj_grad_hess — Method

Negative log-likelihood objective, gradient and hessian estimation using local regressions. Note that this does not ensure positive definiteness of the Hessian.

SyntheticLikelihood.linear_regression — Method

Carry out linear regression. X should have a bias column. Returns tuple (β, ŷ).

linear_regression(X, y)

SyntheticLikelihood.log_prior_gradient — Method

Automatic differentiation to get prior gradient.

SyntheticLikelihood.log_prior_hessian — Method

Automatic differentiation to get prior Hessian.

SyntheticLikelihood.obj_grad_hess — Method

Estimate negative log-likelihood, and its gradient and hessian.

SyntheticLikelihood.obj_grad_hess — Method

Estimate negative log-posterior, and its gradient and hessian.

SyntheticLikelihood.outlier_rows — Method

Find outlier rows in matrix, using threshold multiples of iqr above and below the median for each column.

SyntheticLikelihood.peturb — Method

Peturb a vector using a user specified distribution (often MVN zero mean). Returns array of size (n, length(θ)). If a prior is provided, proposals are checked to have prior support using insupport and are resampled if not.

peturb(θ, d; n)

Arguments

θ Parameter to peturb.

d Distribution from which to sample (see Distributions.jl).

n = 1 Number of peturbed vectors to return.

valid_params Return true if θ vector is valid, and false if invalid.

SyntheticLikelihood.posterior_calc — Method

Get the posterior, gradient and Hessian of negative log-posterior, from the prior and the objective, gradient and hessian of the negative log-likelihood.

SyntheticLikelihood.quadratic_design_matrix — Method

Design matrix for quadratic regression. Bias term appended as first column internally. Returns a tuple, with the matrix and the corresponding indices multiplied, that give rise to each column. Note, indices [1, 1] corresponds to the bias term (so indices compared to original matrix is shifted).

quadratic_design_matrix(X)

SyntheticLikelihood.quadratic_local_μ — Method

Finds the local behaviour of the summary statistic mean μ. Uses quadratic linear regression to approximate the mean, gradient and hessian around θᵢ. Returns a Localμ struct (see above).

quadratic_local_μ(; θᵢ, θ, s)

Arguments

θᵢ::AbstractVector Original θ.

θ::AbstractMatrix Peturbed θ (sampled from local area).

s::AbstractMatrix Corresponding summary statistics to θ.

SyntheticLikelihood.regularize_cor — Method

Regularize correlation matrix, by limiting the condition number c, and putting all correlations below threshold τ to zero.

SyntheticLikelihood.regularize_Σ_merge — Function

Regularize the correlation matrix, by shrinking to the reference to make variances fall within thresholds. lo and hi are multiplied by the variances of the reference to find thresholds. Shrinkage is carried out using Σ = αΣ + (1 - α)ref.

SyntheticLikelihood.remove_invariant — Method

remove columns that have zero variance

remove_invariant(s, s_true)

SyntheticLikelihood.rm_outliers — Method

Remove outlier rows from both matrices, using first matrix to determine outliers.

SyntheticLikelihood.rm_outliers — Method

Remove outlier rows from matrix.

SyntheticLikelihood.simplify_data — Method

Loop through named tuple and call stack_arrays on any vector whose elements are an array. Used at end of samplers.

SyntheticLikelihood.simulate_n_s — Method

As for above, but a Matrix of parameter values are used, carrying out one simulation from each row of θ (and hence n_sim is not required).

simulate_n_s(θ; simulator, summary, parallel)

SyntheticLikelihood.simulate_n_s — Method

Simulates summary statistics from the model under a fixed parameter vector. n_sim is specified as the number of simulations. Simulations can be run on multiple threads using parallel = false. By defualt no summary statistic function is used (by passing the identity function).

simulate_n_s(θ; simulator, summary, n_sim, parallel)

Arguments

θ::AbstractVector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator (defualt identity).

n_sim::Integer Number of simulations.

parallel::Bool = false Whether to run on multiple threads.

SyntheticLikelihood.soft_abs — Method

Takes "soft" absolute value of the eigenvalues, using the method of Betancourt 2013 (https://arxiv.org/pdf/1212.4693.pdf). α → Inf then this approaches the actual absolute value. Minimum eigenvalues are limited to 1/α.

SyntheticLikelihood.stack_arrays — Method

Stacks a vector of consitently sized arrays to make a new array with dimensions (length(x), dim(x[1])...).

stack_arrays(x)

SyntheticLikelihood.standardize — Method

Standardize matrix and vector, using the mean and standard deviation of the matrix.

standardize(X, y)

SyntheticLikelihood.standardize — Method

Standardize to zero mean and standard deviation 1. Can also provide a vector which will be standardized using the mean and standard deviation of the matrix.

standardize(X)