Simulation interface

This section describes the structure a simulator should have to work with the package. Simulator functions should take a single positional argument which is a vector of parameters. Summary functions should take the output from the simulator directly as a single positional argument. The summary function defualts to the identity function, meaning the output of of the simulator is used directly.

The primary function for carrying out simulations is simulate_n_s, which simulates n sets of summary statistics (or the raw simulator output if no summary is specified).

SyntheticLikelihood.simulate_n_s — Function

Simulates summary statistics from the model under a fixed parameter vector. n_sim is specified as the number of simulations. Simulations can be run on multiple threads using parallel = false. By defualt no summary statistic function is used (by passing the identity function).

simulate_n_s(θ; simulator, summary, n_sim, parallel)

Arguments

θ::AbstractVector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator (defualt identity).

n_sim::Integer Number of simulations.

parallel::Bool = false Whether to run on multiple threads.

As for above, but a Matrix of parameter values are used, carrying out one simulation from each row of θ (and hence n_sim is not required).

simulate_n_s(θ; simulator, summary, parallel)