SyntheticLikelihood

Documentation for SyntheticLikelihood package. Simulator functions should have a single positional argument of parameters (and can have arbitrary key word arguments). Summary functions should take the output from the simulator directly as a single positional argument (and can have arbitrary key word arguments).

Function list

SyntheticLikelihood.Localμ

SyntheticLikelihood.get_local_μ

SyntheticLikelihood.linear_regression

SyntheticLikelihood.pairwise_combinations

SyntheticLikelihood.peturb

SyntheticLikelihood.quadratic_transform

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.synthetic_likelihood

Function descriptions

SyntheticLikelihood.Localμ — Type

Struct that contains the information from the first local regression.

SyntheticLikelihood.get_local_μ — Method

Gets the local behaviour of μ. Returns Localμ struct, containing the first and second derivitive estimates as well as the regression residuals.

get_local_μ(; θ_orig, θ, s)

Arguments

θ_orig::AbstractVector Original θ.

θ::AbstractMatrix Peturbed θ (sampled from local area).

s::AbstractMatrix Corresponding summary statistics to θ.

SyntheticLikelihood.linear_regression — Method

Carry out linear regression. X should have a bias column. Returns tuple (β, ŷ).

linear_regression(X, s)

SyntheticLikelihood.pairwise_combinations — Method

Pairwise combinations (for quadratic regression). n=5 would return all the pairwise combinations between 1:5 (including matched terms e.g. [1,1]).

pairwise_combinations(n)

SyntheticLikelihood.peturb — Function

Peturb a vector using a user specified distribution (often MVN zero mean). Returns array of size (n, length(θ))

peturb(θ, d)
peturb(θ, d, n)

Arguments

θ::AbstractVector Parameter to peturb.

d::Sampleable Distribution from which to sample (see Distributions.jl).

n::Int = 1 Number of peturbed vectors to return.

SyntheticLikelihood.quadratic_transform — Method

Quadratic transform. Bias term appended as first column internally. Returns a tuple, with the matrix and the corresponding indices multiplied, that give rise to each column. Note, indices [1, 1] corresponds to the bias term (so indices compared to original matrix is shifted). A bit naive so could be sped up but neat.

quadratic_transform(X)

SyntheticLikelihood.simulate_n_s — Method

Simulates summary statistics from the model under a fixed parameter vector, or a matrix of parameter vectors. If a vector is used, n_sim is specified as the number of simulations. If a matrix is used, one simulation is carried out using each row as a parameter vector. By defualt simulations are run on multiple threads (see Threads manual for information on starting Julia with multiple threads).

simulate_n_s(θ; simulator, summary, n_sim, simulator_kwargs, summary_kwargs, parallel)

Arguments

θ::AbstractVector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator.

n_sim::Int Number of simulations.

simulator_kwargs Kwargs passed to simulator.

summary_kwargs Kwargs passed to summary.

parallel::Bool = true Whether to run on multiple threads.

SyntheticLikelihood.simulate_n_s — Method

As for above, but uses an Array of parameter values are used, carrying out one simulation from each row of θ.

simulate_n_s(θ; simulator, summary, simulator_kwargs, summary_kwargs, parallel)

SyntheticLikelihood.synthetic_likelihood — Method

Evaluates synthetic likelhood of observed data for a fixed parameter vector using a multivariate Gaussian assumption as in (Simon Wood, 2010).

synthetic_likelihood(θ; simulator, summary, s_true, n_sim, simulator_kwargs, summary_kwargs)

Arguments

θ::AbstractVector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator.

s_true::AbstractVector Observed summary statistics.

n_sim::Int Number of simulations to use.

simulator_kwargs Kwargs splatted in simulator.

summary_kwargs Kwargs splatted in summary.