SyntheticLikelihood

Documentation for SyntheticLikelihood package. Simulator functions should take a single positional argument which is a vector of parameters (and can have arbitrary key word arguments). Summary functions should take the output from the simulator directly as a single positional argument (and can have arbitrary key word arguments).

Function list

SyntheticLikelihood.AbstractSamplerState

SyntheticLikelihood.BasicState

SyntheticLikelihood.GradientHessianState

SyntheticLikelihood.GradientState

SyntheticLikelihood.LocalΣ

SyntheticLikelihood.Localμ

SyntheticLikelihood.LangevinDiffusion

SyntheticLikelihood.add_state!

SyntheticLikelihood.get_residuals

SyntheticLikelihood.glm_local_Σ

SyntheticLikelihood.init_data_tuple

SyntheticLikelihood.linear_regression

SyntheticLikelihood.pairwise_combinations

SyntheticLikelihood.peturb

SyntheticLikelihood.quadratic_design_matrix

SyntheticLikelihood.quadratic_local_μ

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.synthetic_likelihood

Function descriptions

SyntheticLikelihood.AbstractSamplerState — Type

Concrete sampler types should have fields:

θ::AbstractVector

objective::Float64

counter::Int

SyntheticLikelihood.BasicState — Type

Struct for containing the state of sampler at each iteration for simple samplers (that just require objective evaluation).

SyntheticLikelihood.GradientHessianState — Type

Struct for containing the state of sampler at each iteration for samplers which use the gradient and hessian.

SyntheticLikelihood.GradientState — Type

Struct for containing the state of sampler at each iteration for simple gradient based samplers.

SyntheticLikelihood.LocalΣ — Type

Struct that contains the estimated local properties of Σ (the covariance matrix of the summary statistics).

Fields

Σ The expected value of the diagonal covariance matrix (of summary statistics). A vector of length nsumstats.

∂ The first derivitives of Σ, of dimension (ns × ns × n_θ).

SyntheticLikelihood.Localμ — Type

Struct that contains the estimated local properties of μ.

Fields

μ::Float64 Mean of the summary statistic.

∂::Vector{Float64} First derivitive w.r.t. parameters.

∂²::Matrix{Float64} Second derivitive w.r.t. parameters.

ϵ::Vector{Float64} Residuals.

SyntheticLikelihood.LangevinDiffusion — Method

Sample using Langevin diffusion (unadjusted Langevin algorithm). This uses the update θ := θ + step_size/2 .* ∇θ .+ ξ. ξ is usually Brownian noise. Uses a fixed step size.

Returns a tuple, (data, state), where data is a NamedTuple of stored data, and state is the state at the final iteration.

Arguments: state::GradientState Initial starting state for sampler. objective::Function Objective function (assumed aim would be to maximize) gradient::Function Gradient of the objective function with respect to the parameters. step_size Multiplied elementwise by gradient. ξ::Sampleable Distribution to add noise to the diffusion. Added elementwise. n_steps::Int Number of iterations to carry out. data_to_collect::AbstractVector{Symbol} Vector of symbols, denoting the items in the state to store at each iteration.

LangevinDiffusion(; state, objective, gradient, step_size, ξ, n_steps, data_to_collect)

SyntheticLikelihood.add_state! — Method

Add data to the data tuple.

SyntheticLikelihood.get_residuals — Method

Get an array of residuals from a vector of Localμ structs. Returns (length of residuals × number of summary stats) matrix

SyntheticLikelihood.glm_local_Σ — Method

Use a gamma distributed GLM with log link function to estimate the local properties of Σ. θ should not have a bias term (added internally).

Arguments

θ_orig::AbstractVector Original θ (used for centering).

θ::AbstractMatrix Peturbed θ from local area.

ϵ::AbstractMatrix Residuals from quadratic regression.

SyntheticLikelihood.init_data_tuple — Method

Function initialises a named tuple containing Vectors with undefined values. Used with samplers to store results. State just provides an "example" state from which to infer types of vectors in the array. Names of the named tuple are the symbols provided.

SyntheticLikelihood.linear_regression — Method

Carry out linear regression. X should have a bias column. Returns tuple (β, ŷ).

linear_regression(X, y)

SyntheticLikelihood.pairwise_combinations — Method

Pairwise combinations (for quadratic regression). n=5 would return all the pairwise combinations between 1:5 (including matched terms e.g. [1,1]).

pairwise_combinations(n)

SyntheticLikelihood.peturb — Function

Peturb a vector using a user specified distribution (often MVN zero mean). Returns array of size (n, length(θ))

peturb(θ, d)
peturb(θ, d, n)

Arguments

θ::AbstractVector Parameter to peturb.

d::Sampleable Distribution from which to sample (see Distributions.jl).

n::Int = 1 Number of peturbed vectors to return.

SyntheticLikelihood.quadratic_design_matrix — Method

Design matrix for quadratic regression. Bias term appended as first column internally. Returns a tuple, with the matrix and the corresponding indices multiplied, that give rise to each column. Note, indices [1, 1] corresponds to the bias term (so indices compared to original matrix is shifted).

quadratic_design_matrix(X)

SyntheticLikelihood.quadratic_local_μ — Method

Finds the local behaviour of the summary statistic mean μ. Uses quadratic linear regression to approximate the mean, gradient and hessian around θ_orig. Returns a vector of Localμ structs (see above), with length equal to the number of summary statistics.

quadratic_local_μ(; θ_orig, θ, s)

Arguments

θ_orig::AbstractVector Original θ.

θ::AbstractMatrix Peturbed θ (sampled from local area).

s::AbstractMatrix Corresponding summary statistics to θ.

SyntheticLikelihood.simulate_n_s — Method

Simulates summary statistics from the model under a fixed parameter vector. n_sim is specified as the number of simulations. Simulations can be run on multiple threads using parallel = true. By defualt no summary statistic function is used (by passing the identity function).

simulate_n_s(θ; simulator, summary, n_sim, simulator_kwargs, summary_kwargs, parallel)

Arguments

θ::AbstractVector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator (defualt identity).

n_sim::Int Number of simulations.

simulator_kwargs Kwargs passed to simulator.

summary_kwargs Kwargs passed to summary.

parallel::Bool = true Whether to run on multiple threads.

SyntheticLikelihood.simulate_n_s — Method

As for above, but a Matrix of parameter values are used, carrying out one simulation from each row of θ (and hence n_sim is not required).

simulate_n_s(θ; simulator, summary, simulator_kwargs, summary_kwargs, parallel)

SyntheticLikelihood.synthetic_likelihood — Method

Evaluates synthetic likelhood of observed data for a fixed parameter vector using a multivariate Gaussian assumption as in (Simon Wood, 2010).

synthetic_likelihood(θ; simulator, summary, s_true, n_sim, simulator_kwargs, summary_kwargs)

Arguments

θ::AbstractVector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator (defualt identity).

s_true::AbstractVector Observed summary statistics.

n_sim::Int Number of simulations to use.

simulator_kwargs Kwargs splatted in simulator.

summary_kwargs Kwargs splatted in summary.