SyntheticLikelihood

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.synthetic_likelihood

SyntheticLikelihood.simulate_n_s — Method

Simulates summary statistics from a matrix of parameter vectors.

Arguments

θ::Array Array of parameters. Each row is passed to simulator.

simulator::Function Simulator function.

summary::Function Summary function that takes output of simulator.

simulator_kwargs Kwargs passed to simulator.

summary_kwargs Kwargs passed to summary.

SyntheticLikelihood.simulate_n_s — Method

Simulates summary statistics from the model under a fixed parameter vector.

Arguments

θ::Vector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator.

n_sim::Int Number of simulations.

simulator_kwargs Kwargs passed to simulator.

summary_kwargs Kwargs passed to summary.

SyntheticLikelihood.synthetic_likelihood — Method

Evaluates a point estimate of the synthetic likelihood.

Arguments

θ::Vector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator.

s_true::Vector Observed summary statistics.

n_sim::Int Number of simulations to use.

simulator_kwargs Kwargs passed to simulator.

summary_kwargs Kwargs passed to summary.