SyntheticLikelihood

Documentation for SyntheticLikelihood package. The package is currently a work in progress. Simulator functions should take a single positional argument which is a vector of parameters (and can have arbitrary key word arguments). Summary functions should take the output from the simulator directly as a single positional argument (and can have arbitrary key word arguments).

Examples

Sampling

Implementation details

Function list

SyntheticLikelihood.Langevin

SyntheticLikelihood.LocalΣ

SyntheticLikelihood.Localμ

SyntheticLikelihood.PreconditionedLangevin

SyntheticLikelihood.SamplerState

SyntheticLikelihood.add_state!

SyntheticLikelihood.get_residuals

SyntheticLikelihood.glm_local_Σ

SyntheticLikelihood.init_data_tuple

SyntheticLikelihood.linear_regression

SyntheticLikelihood.pairwise_combinations

SyntheticLikelihood.peturb

SyntheticLikelihood.quadratic_design_matrix

SyntheticLikelihood.quadratic_local_μ

SyntheticLikelihood.run_sampler!

SyntheticLikelihood.simplify_data

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.simulate_n_s

SyntheticLikelihood.stack_arrays

SyntheticLikelihood.synthetic_likelihood

Function descriptions

SyntheticLikelihood.Langevin — Type

Sampler object for Langevin diffusion. Uses a discrete time Euler approximation of the Langevin diffusion (unadjusted Langevin algorithm), given by the update θ := θ - η/2 * ∇ + ξ, where ξ is given by N(0, ηI).

SyntheticLikelihood.LocalΣ — Type

Struct that contains the estimated local properties of Σ (the covariance matrix of the summary statistics).

Fields

Σ The (nₛ×nₛ) (estimated) covariance matrix of the summary statistics.

∂ The (nₛ×nₛ×n_θ) matrix of estimated first derivitives of Σ.

SyntheticLikelihood.Localμ — Type

Struct that contains the estimated local properties of μ.

Fields

μ::Float64 Mean of the summary statistic.

∂::Vector{Float64} First derivitive w.r.t. parameters.

∂²::Matrix{Float64} Second derivitive w.r.t. parameters.

ϵ::Vector{Float64} Residuals.

SyntheticLikelihood.PreconditionedLangevin — Type

Sampler object for Preconditioned Langevin diffusion. Also can be thought of as a stochastic newton method. Uses the update: θ := θ - η/2 * H⁻¹*∇ + ξ, where ξ ∼ N(0, ηH⁻¹).

SyntheticLikelihood.SamplerState — Type

Struct for containing the state of sampler at a particular iteration. Gradient and hessian are missing unless specified.

SyntheticLikelihood.add_state! — Method

Add data to the data tuple.

SyntheticLikelihood.get_residuals — Method

Get an array of residuals from a vector of Localμ structs. Returns (length of residuals × number of summary stats) matrix

SyntheticLikelihood.glm_local_Σ — Method

Use a gamma distributed GLM with log link function to estimate the local properties of Σ. θ should not have a bias term (added internally).

Arguments

θ_orig::AbstractVector Original θ (used for centering).

θ::AbstractMatrix Peturbed θ from local area.

ϵ::AbstractMatrix Residuals from quadratic regression.

SyntheticLikelihood.init_data_tuple — Method

Function initialises a named tuple containing Vectors with undefined values. Used with samplers to store results. State just provides an "example" state from which to infer types of vectors in the array. Names of the named tuple are the symbols provided.

SyntheticLikelihood.linear_regression — Method

Carry out linear regression. X should have a bias column. Returns tuple (β, ŷ).

linear_regression(X, y)

SyntheticLikelihood.pairwise_combinations — Method

Pairwise combinations (for quadratic regression). n=5 would return all the pairwise combinations between 1:5 (including matched terms e.g. [1,1]).

pairwise_combinations(n)

SyntheticLikelihood.peturb — Function

Peturb a vector using a user specified distribution (often MVN zero mean). Returns array of size (n, length(θ))

peturb(θ, d)
peturb(θ, d, n)

Arguments

θ::AbstractVector Parameter to peturb.

d::Sampleable Distribution from which to sample (see Distributions.jl).

n::Integer = 1 Number of peturbed vectors to return.

SyntheticLikelihood.quadratic_design_matrix — Method

Design matrix for quadratic regression. Bias term appended as first column internally. Returns a tuple, with the matrix and the corresponding indices multiplied, that give rise to each column. Note, indices [1, 1] corresponds to the bias term (so indices compared to original matrix is shifted).

quadratic_design_matrix(X)

SyntheticLikelihood.quadratic_local_μ — Method

Finds the local behaviour of the summary statistic mean μ. Uses quadratic linear regression to approximate the mean, gradient and hessian around θ_orig. Returns a vector of Localμ structs (see above), with length equal to the number of summary statistics.

quadratic_local_μ(; θ_orig, θ, s)

Arguments

θ_orig::AbstractVector Original θ.

θ::AbstractMatrix Peturbed θ (sampled from local area).

s::AbstractMatrix Corresponding summary statistics to θ.

SyntheticLikelihood.run_sampler! — Function

Run the sampling algorithm. Data to collect at each iteration is specified by collect_data, and should be a subset of [:θ, :objective, :gradient, :hessian, :counter].

run_sampler!(sampler, init_θ, n_steps)
run_sampler!(sampler, init_θ, n_steps, collect_data)

Returns a tuple, with keys matching collect_data.

SyntheticLikelihood.simplify_data — Method

Loop through named tuple and call stack_arrays on any vector whose elements are an array. Used at end of samplers.

SyntheticLikelihood.simulate_n_s — Method

Simulates summary statistics from the model under a fixed parameter vector. n_sim is specified as the number of simulations. Simulations can be run on multiple threads using parallel = true. By defualt no summary statistic function is used (by passing the identity function).

simulate_n_s(θ; simulator, summary, n_sim, simulator_kwargs, summary_kwargs, parallel)

Arguments

θ::AbstractVector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator (defualt identity).

n_sim::Integer Number of simulations.

simulator_kwargs Kwargs passed to simulator.

summary_kwargs Kwargs passed to summary.

parallel::Bool = true Whether to run on multiple threads.

SyntheticLikelihood.simulate_n_s — Method

As for above, but a Matrix of parameter values are used, carrying out one simulation from each row of θ (and hence n_sim is not required).

simulate_n_s(θ; simulator, summary, simulator_kwargs, summary_kwargs, parallel)

SyntheticLikelihood.stack_arrays — Method

Stacks a vector of consitently sized arrays to make a new array with dimensions (length(x), dim(x[1])...).

SyntheticLikelihood.synthetic_likelihood — Method

Evaluates synthetic likelhood of observed data for a fixed parameter vector using a multivariate Gaussian assumption as in (Simon Wood, 2010).

synthetic_likelihood(θ; simulator, summary, s_true, n_sim, simulator_kwargs, summary_kwargs)

Arguments

θ::AbstractVector Parameter vector passed to simulator.

simulator::Function Simulator.

summary::Function Summary function that takes output of simulator (defualt identity).

s_true::AbstractVector Observed summary statistics.

n_sim::Integer Number of simulations to use.

simulator_kwargs Kwargs splatted in simulator.

summary_kwargs Kwargs splatted in summary.