Local regression

Local regressions can be used to estimate the gradient and hessian of the likelihood using local_synthetic_likelihood.

Below is an example, inferring the means of a 10 dimensional multivariate normal distribution (with constant covariance), using a single observed sample from the distribution.

using SyntheticLikelihood, Distributions, Plots

# Define the simulator
function simulator(θ::Vector{Float64})
  @assert length(θ) == 10
  d = MvNormal(θ, sqrt(0.1))
  rand(d)
end

# Generate pseudo-observed data
θ_true = zeros(10)
s_true = simulator(θ_true)

10-element Vector{Float64}:
  0.06448714181467803
  0.970688236987571
 -0.0047472257692524585
  0.4282728511476993
  0.04330026361454093
  0.014661330742861271
 -0.14120854118328582
 -0.3075131435136663
  0.636819168467455
 -0.08537114746259376

We can then define how to sample from the distribution. Below I will use the PreconditionedLangevin sampler with a step size of 0.1, and use local_synthetic_likelihood to estimate the gradient and Hessian of the likelihood. The hyperparameters to be used with local_synthetic_likelihood are passed as key word arguments to the sampler:

P = MvNormal(fill(0.5, 10))  # Local area in parameter space
n_sim = 1000  # Simulations used at each iteration
θ_orig = convert(Vector{Float64}, -5:5)  # Starting parameter values.

pl = PreconditionedLangevin(
  0.1, local_synthetic_likelihood; s_true, simulator, P, n_sim
  )

PreconditionedLangevin(0.1, SyntheticLikelihood.local_synthetic_likelihood, Base.Iterators.Pairs{Symbol, Any, NTuple{4, Symbol}, NamedTuple{(:s_true, :simulator, :P, :n_sim), Tuple{Vector{Float64}, typeof(Main.ex-1.simulator), Distributions.ZeroMeanDiagNormal{Tuple{Base.OneTo{Int64}}}, Int64}}}(:s_true => [0.06448714181467803, 0.970688236987571, -0.0047472257692524585, 0.4282728511476993, 0.04330026361454093, 0.014661330742861271, -0.14120854118328582, -0.3075131435136663, 0.636819168467455, -0.08537114746259376], :simulator => Main.ex-1.simulator, :P => ZeroMeanDiagNormal(
dim: 10
μ: 10-element Zeros{Float64}
Σ: [0.25 0.0 … 0.0 0.0; 0.0 0.25 … 0.0 0.0; … ; 0.0 0.0 … 0.25 0.0; 0.0 0.0 … 0.0 0.25]
)
, :n_sim => 1000))

Now we can cary out the sampling and plot the results.

data = run_sampler!(pl, θ_orig, 1000)

# Plot the samples
param_names = reshape(["θ$i" for i in 1:10], (1,10))
plot(data.θ, label = param_names)

Plotting the marginals after removing the burn in period:

histogram(data.θ[250:end, :], label = param_names, layout = (2,5), size = (800, 600))

We can see the samples are generally centered around the true parameter values (all zeros). More specifically, they are centered around s_true in this case, which are generally around zero.

SyntheticLikelihood.local_synthetic_likelihood — Function

Estimate negative log-synthetic likelihood, and its gradient and hessian.

local_synthetic_likelihood(θ; s_true, simulator, summary, P, n_sim, eigval_threshold)

Arguments

θ Starting parameter values.

P::Sampleable Distribution used to peturb parameter value (e.g. 0 mean multivariate normal).

s_true The observed summary statistics.

simulator The simulator function taking parameter vector θ.

summary The summary function taking output from the simulator.

n_sim The number of peturbed points to use for the local regression.

eigval_threshold Minimum eigenvalue threshold for the estimated hessian.