Local regression

Local regressions can be used to estimate the gradient and hessian of the likelihood, which can be used to improve the sampling efficieny of synthetic likelihood.

Example

Here we consider a simple example, in which we infer the mean of a 10-dimensional multivariate normal distribution, using simulations from the distribution.

Define the simulator

The simulator must take a single positional argument, which is the parameter vector:

using SyntheticLikelihood, Distributions, Plots

# Define the simulator
function simulator(θ::Vector{Float64})
  @assert length(θ) == 10
  d = MvNormal(θ, sqrt(0.1))
  rand(d)
end

Ground truth

The "true" parameters which we will aim to estimate, is just a vector of zeros. We can use this to generate a pseudo-observed data set s_true.

θ_true = zeros(10)
s_true = simulator(θ_true)

10-element Vector{Float64}:
  0.7337503594175706
  0.10492845444800385
  0.3558951997281131
 -0.13685191084987683
  0.46090029496657503
 -0.17799035666713817
  0.028367106124074244
 -0.14246306174905
  0.1284909623908065
 -0.683900249687327

Defining how to estimate the likelihood

We can then define the hyperparameters for estimating the likelihood using local regression using LocalLikelihood.

local_likelihood = LocalLikelihood(;
  simulator, s_true,
  P = MvNormal(fill(0.5, 10)),
  n_sim = 1000
)

Note that if required a summary function can optionally be specified here, to summarise the output of the simulator.

Defining sampling method

We can then define how to sample from the distribution. Below I will use the RiemannianULA sampler with a step size of 0.1.

pULA = RiemannianULA(0.1)

Sampling

We can now define some initial parameter values, init_θ, and sample from the distribution using run_sampler!:

init_θ = convert(Vector{Float64}, 1:10)
data = run_sampler!(pULA, local_likelihood; init_θ, n_steps = 500)

Plot the samples

param_names = reshape(["θ$i" for i in 1:10], (1,10))
plot(data.θ, label = param_names)

We can see that after the burn in period, samples are generally centered around the true parameter values (all zeros). More specifically, they are centered around s_true in this case, which are generally around zero.

Bayesian inference

Given a prior, it is also simple to sample from the posterior instead of the likelihood. The prior should be specified using the distributions from Distributions.jl. A multivariate distribution can be used, or alternatively the prior can be formed from independent univariate priors using a Product distribution from Distributions.jl. For this example the prior is a multivariate normal centered around 5, with no correlation structure, and σ=0.5:

prior = MvNormal(fill(5, 10), 0.5)

We can then define our objective using LocalPosterior and run the sampler again:

local_posterior = LocalPosterior(prior, local_likelihood)
data = run_sampler!(pULA, local_posterior; init_θ, n_steps = 500)
plot(data.θ, label = param_names)

Note internally, this uses LocalLikelihood to estimate the the gradient and Hessian of the likelihood as before, and then uses automatic differentiation of the prior to get the gradient and Hessian of the prior. These can then be used to calculcate the gradient and Hessian of the posterior.

Currently available "objectives"

SyntheticLikelihood.LocalLikelihood — Type

Contains the hyperparameters for getting a local approximation of the negative log-likelihood surface using local regressions.

simulator

summary

s_true

P

Initial distribution used to peturb the parameter value.

n_sim

The number of peturbed points to use for the local regression.

P_regularizer

Adaptive proposal distribution. Should not be set manually.

SyntheticLikelihood.LocalPosterior — Type

Contains the hyperparameters for getting a local approximation of the posterior.

prior

simulator

summary

s_true

P

n_sim

P_regularizer