Local regression

Local regressions can be used to estimate the gradient and hessian of the likelihood using local_synthetic_likelihood.

SyntheticLikelihood.local_synthetic_likelihood — Function

Estimate negative log-synthetic likelihood, and its gradient and hessian.

local_synthetic_likelihood(θ; s_true, simulator, summary, P, n_sim, eigval_threshold)

Arguments

θ Starting parameter values.

P Distribution used to peturb parameter value (e.g. 0 mean multivariate normal).

s_true The observed summary statistics.

simulator The simulator function taking parameter vector θ.

summary The summary function taking output from the simulator.

n_sim The number of peturbed points to use for the local regression.

eigval_threshold = 0.5 Minimum eigenvalue threshold for the estimated hessian. Negative eigenvalues are flipped and those smaller than the threshold are set to the threshold.

Below is an example, inferring the means of a 10 dimensional multivariate normal distribution (with constant covariance), using a single observed sample from the distribution.

using SyntheticLikelihood, Distributions, Plots

# Define the simulator
function simulator(θ::Vector{Float64})
  @assert length(θ) == 10
  d = MvNormal(θ, sqrt(0.1))
  rand(d)
end

# Generate pseudo-observed data
θ_true = zeros(10)
s_true = simulator(θ_true)

10-element Vector{Float64}:
 -0.062395850769133836
  0.18389180958611182
  0.06114078604218306
  0.08501852125089139
  0.43350678212397326
 -0.1835121555079424
  0.697453853960038
 -0.5355072489188741
  0.4767033597993553
 -0.17584737380728344

We can then define how to sample from the distribution. Below I will use the PreconditionedLangevin sampler with a step size of 0.1, and use local_synthetic_likelihood to estimate the gradient and Hessian of the likelihood. The hyperparameters to be used with local_synthetic_likelihood are passed as key word arguments to the sampler:

P = MvNormal(fill(0.5, 10))  # Local area in parameter space
n_sim = 1000  # Simulations used at each iteration
θ_orig = convert(Vector{Float64}, 1:10)  # Starting parameter values.

pl = PreconditionedLangevin(
  0.1, local_synthetic_likelihood; s_true, simulator, P, n_sim
  )

# Carry out the sampling and plot the results
data = run_sampler!(pl, θ_orig, 1000)

# Plot the samples
param_names = reshape(["θ$i" for i in 1:10], (1,10))
plot(data.θ, label = param_names)

Plotting the marginals after removing the burn in period:

histogram(data.θ[250:end, :], label = param_names, layout = (2,5), size = (800, 600))

We can see the samples are generally centered around the true parameter values (all zeros). More specifically, they are centered around s_true in this case, which are generally around zero.