Samplers

To sample from a distribution, first a sampler object should be created. The currently available samplers are shown below:

SyntheticLikelihood.Langevin — Type

Sampler object for Langevin diffusion. Uses a discrete time Euler approximation of the Langevin diffusion (unadjusted Langevin algorithm), given by the update θ := θ - η/2 * ∇ + ξ, where ξ is given by N(0, ηI).

step_size

local_approximation

Step size parameter.

kwargs

Must return LocalApproximation object with objective and gradient fields.

SyntheticLikelihood.PreconditionedLangevin — Type

Sampler object for Preconditioned Langevin diffusion. Also can be thought of as a stochastic newton method. Uses the update: θ := θ - η/2 * H⁻¹*∇ + ξ, where ξ ∼ N(0, ηH⁻¹).

step_size

local_approximation

Step size parameter

kwargs

Must return LocalApproximation object with objective, gradient and hessian.

The sampler object defines the hyperparameters of the sampler, and a function local_approximation, which takes θ and returns a LocalApproximation object, with fields objective, gradient and hessian. The gradient and hessian defualt to nothing if not required. This approach seems a bit convoluted (compared to e.g. seperately passing objective, gradient and Hessian functions), but it facilitates reusing calculations shared between calculating the objective, gradient and hessian, if desired. The aim should be to explore around the minima of the function, so the objective could be the negative log-posterior, for example.

The sampler object can then passed to run_sampler!, to sample from the distribution:

SyntheticLikelihood.run_sampler! — Function

Run the sampling algorithm. Data to collect at each iteration is specified by collect_data, and should be a subset of [:θ, :objective, :gradient, :hessian, :counter].

run_sampler!(sampler, init_θ, n_steps)
run_sampler!(sampler, init_θ, n_steps, collect_data)

Returns a tuple, with keys matching collect_data.

Below is an example to sample from a multivariate normal density using the discretized langevin diffusion (Unadjusted Langevin Algorithm).

using SyntheticLikelihood, Distributions, Plots

# Sample from MVN
d = MvNormal([10 5; 5 10])
function local_approximation(θ)
    LocalApproximation(objective = -logpdf(d, θ),
                       gradient = -gradlogpdf(d, θ))
end

init_θ = [-15., -15]
n_steps = 1000

langevin = Langevin(1., local_approximation)
data = run_sampler!(langevin, init_θ, n_steps, [:θ, :counter])

θ_samples = data[:θ]

# Plot samples
x = y = range(-20, 20; length=50)
f(x, y) = -logpdf(d, [x,y])
X = repeat(reshape(x, 1, :), length(y), 1)
Y = repeat(y, 1, length(x))
Z = map(f, X, Y)
p = contour(x, y, f)
scatter!(θ_samples[:,1], θ_samples[:,2], legend = false)

Note that here a determinisic local_approximation function is used. Generally, in simulation-based-inference this would not be available, and hence this can be replaced for example with local_synthetic_likelihood, which uses local regressions to approximate the gradient and Hessian of the likelihood function.

Implementation details

To implement a new sampler, each sampler must:

Be a subtype of AbstractSampler.

Have fields local_approximation and kwargs (that are passed to local_approximation).

Have a get_init_state method, which returns a SamplerState object given init_θ e.g. with signature get_init_state(sampler::MySampler, init_θ::Vector{Float64}).

Have an update! method, taking the sampler and the SamplerState object, e.g. update!(sampler::MySampler, state::SamplerState). This updates the state (parameters, gradients objective function value etc, and sampler object if applicable).