Samplers

To sample from a distribution, first a sampler object should be created.

The currently available samplers are shown below:

SyntheticLikelihood.Langevin — Type

Sampler object for Langevin diffusion. Uses a discrete time Euler approximation of the Langevin diffusion (unadjusted Langevin algorithm), given by the update θ := θ - η/2 * ∇ + ξ, where ξ is given by N(0, ηI).

step_size

objective

gradient

SyntheticLikelihood.PreconditionedLangevin — Type

Sampler object for Preconditioned Langevin diffusion. Also can be thought of as a stochastic newton method. Uses the update: θ := θ - η/2 * H⁻¹*∇ + ξ, where ξ ∼ N(0, ηH⁻¹).

step_size

objective

gradient_hessian

The sampler object defines the hyperparameters of the sampler, the objective function and the gradient/hessian if appropriate. When passing a function (objective/gradient/hessian) to a sampler, it should take a parameter vector θ as its first and only required argument. The aim should be to explore around the minima of the function, so the objective could be the negative log-posterior, for example.

The sampler object can then passed to run_sampler!, to sample from the distribution:

SyntheticLikelihood.run_sampler! — Function

Run the sampling algorithm. Data to collect at each iteration is specified by collect_data, and should be a subset of [:θ, :objective, :gradient, :hessian, :counter].

run_sampler!(sampler, init_θ, n_steps)
run_sampler!(sampler, init_θ, n_steps, collect_data)

Returns a tuple, with keys matching collect_data.

Below is an example to sample from a multivariate normal density using the discretized langevin diffusion (Unadjusted Langevin Algorithm).

using SyntheticLikelihood, Distributions, Gadfly

# Sample from MVN
d = MvNormal([10 5; 5 10])
objective(θ) = -logpdf(d, θ)
gradient(θ) = -gradlogpdf(d, θ)
init_θ = [-15., -15]
n_steps = 1000

langevin = Langevin(;step_size = 1., objective, gradient)
data = run_sampler!(langevin, init_θ, n_steps, [:θ, :counter])

θ_samples = data[:θ]
x = y = range(-20, 20; length=50)

contours = layer(z = (x, y) -> logpdf(d, [x,y]), x = x, y = y, Geom.contour)
points = layer(x = θ_samples[:,1], y = θ_samples[:,2], Geom.point,
                Theme(alphas=[0.2], default_color = "black"))

plot(contours, points)

Implementation details

To implement a new sampler, each sampler must:

Be a subtype of AbstractSampler.

Have a get_init_state method, which returns a SamplerState object given init_θ e.g. with signature get_init_state(sampler::MySampler, init_θ::Vector{Float64}).

Have an update! method, taking the sampler and the SamplerState object update!(sampler::MySampler, state::SamplerState). This updates the state (parameters, gradients etc, and sampler object if applicable).