Samplers

To sample from a distribution, first a sampler object should be created. The currently available samplers are shown below:

SyntheticLikelihood.ULA — Type

Sampler for unadjusted langevin algorithm. Uses a discrete time Euler approximation of the langevin diffusion, given by the update θ := θ - η/2 * ∇ + ξ, where ξ is given by N(0, ηI).

step_size

SyntheticLikelihood.RiemannianULA — Type

Sampler object for Riemannian ULA. Uses the update: θ := θ - ϵ²H⁻¹*∇ - ϵ√H⁻¹ z, where z ∼ N(0, I).

step_size

The sampler object defines the hyperparameters of the sampler, and a function obj_grad_hess, which takes θ and returns a ObjGradHess object, with fields objective, gradient and hessian. The gradient and hessian defualt to nothing if not required. This approach seems a bit convoluted (compared to e.g. seperately passing objective, gradient and Hessian functions), but it facilitates reusing calculations shared between calculating the objective, gradient and hessian, if desired. The aim should be to explore around the minima of the function, so the objective could be the negative log-posterior, for example.

The sampler object can then passed to run_sampler!, to sample from the distribution:

SyntheticLikelihood.run_sampler! — Function

Run the sampling algorithm. Data to collect at each iteration is specified by collect_data, and should be a subset of [:θ, :objective, :gradient, :hessian, :counter].

run_sampler!(sampler, local_approximation; init_θ, n_steps, collect_data)

Returns a tuple, with keys matching collect_data.

Below is an example to sample from a multivariate normal density using the discretized ULA diffusion (Unadjusted ULA Algorithm). In the below example a summary function is not used (it is left to defualt to the identity), so inference is performed on the raw simulator output.

using SyntheticLikelihood, Distributions, Plots

# Sample from MVN
d = MvNormal([10 5; 5 10])
function obj_grad_hess(θ)
    ObjGradHess(objective = -logpdf(d, θ),
                       gradient = -gradlogpdf(d, θ))
end

init_θ = [-15., -15]
n_steps = 1000

ULA = ULA(1., obj_grad_hess)
data = run_sampler!(ULA, init_θ, n_steps, [:θ, :counter])

θ_samples = data[:θ]

# Plot samples
x = y = range(-20, 20; length=50)
f(x, y) = -logpdf(d, [x,y])
X = repeat(reshape(x, 1, :), length(y), 1)
Y = repeat(y, 1, length(x))
Z = map(f, X, Y)
p = contour(x, y, f)
scatter!(θ_samples[:,1], θ_samples[:,2], legend = false)

Note that here a determinisic obj_grad_hess function is used. Generally, in simulation-based-inference this would not be available, and hence this can be replaced for example with local_likelihood, which uses local regressions to approximate the gradient and Hessian of the likelihood function.

Implementation details

To implement a new sampler, each sampler must:

Be a subtype of AbstractSampler.

Have fields containing the hyperparameters.

Have an update! method, taking the sampler the LocalApproximation and the SamplerState object as arguments e.g. update!(sampler::MySampler, local_approximation::LocalApproximation, state::SamplerState). This updates the state (parameters, gradients objective function value etc, and sampler object if applicable).